src/main/java/hackerrank/algorithm/easy/BeautifulTriplets.java
package hackerrank.algorithm.easy;

import java.io.BufferedWriter;
import java.io.FileWriter;
import java.io.IOException;
import java.util.Scanner;

public class BeautifulTriplets {

    // Complete the beautifulTriplets function below.
    static int beautifulTriplets(int d, int[] arr) {
        int sum = 0;
        int[] numberCountArray = getNumberCount(arr);
        int maxIValue = arr[arr.length-1]-d-d;
        for (int i = 0; i <= maxIValue; i++) {
            sum = sum + (numberCountArray[i] * numberCountArray[i+d] * numberCountArray[i+d+d]);
        }
        return sum;
    }

    static int[] getNumberCount(int[] arr) {
        int[] result = new int[arr[arr.length-1]+1];
        for (int i = 0; i < arr.length; i++) {
            result[arr[i]] += 1;
        }
        return result;
    }

    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) throws IOException {
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        String[] nd = scanner.nextLine().split(" ");

        int n = Integer.parseInt(nd[0]);

        int d = Integer.parseInt(nd[1]);

        int[] arr = new int[n];

        String[] arrItems = scanner.nextLine().split(" ");
        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");

        for (int i = 0; i < n; i++) {
            int arrItem = Integer.parseInt(arrItems[i]);
            arr[i] = arrItem;
        }

        int result = beautifulTriplets(d, arr);

        bufferedWriter.write(String.valueOf(result));
        bufferedWriter.newLine();

        bufferedWriter.close();

        scanner.close();
    }
}
